fix: refactor ContactBar
