+++ src/mui/blocks/ContactBar/ContactBar.module.scss
@@ -1,0 +1,49 @@
+@import "styles/modules/all";
+
+.contactBar {
+  width: 100%;
+  height: auto;
+  display: flex;
+
+
+  .inner {
+    display: flex;
+    width: 100%;
+    background-color: var(--mui-palette-TB-Brand-Primary, #0C223F);
+    padding: 20px 5% 70px 5%;
+    flex-direction: column;
+
+    @include breakpoint(xs) {
+      flex-direction: row;
+    }
+
+    .alpha {
+      display: flex;
+      justify-content: center;
+      flex-basis: 100%;
+
+      @include breakpoint(sm) {
+       flex-basis: 50%;
+      }
+      .alphaContainer {
+        display: flex;
+        flex-direction: column;
+        .alphaInner {
+          display: flex;
+          align-items: center;
+          justify-content: flex-start;
+          gap: 1rem;
+          color: var(--mui-palette-secondary-contrastText, #fff);
+
+          img {
+            width: 40px;
+          }
+
+          .title {
+            font-weight: 600;
+          }
+        }
+      }
+    }
+  }
+}
