--- src/storybook/ClassTable.js
@@ -1,33 +1,0 @@
-import * as React from 'react';
-import Paper from '@mui/material/Paper';
-import Table from '@mui/material/Table';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import TableCell from '@mui/material/TableCell';
-import TableBody from '@mui/material/TableBody';
-import TableContainer from '@mui/material/TableContainer';
-
-const ClassTable = ({ rows, title }) => {
-  return (
-    <TableContainer component={Paper}>
-      <Table stickyHeader sx={{ minWidth: 650 }} aria-label='MUI class table'>
-        <TableHead>
-          <TableRow>
-            <TableCell>{title}</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {rows.map((row) => (
-            <TableRow key={row.name} sx={{ '&:last-child td, &:last-child th': { border: 0 } }}>
-              <TableCell component='th' scope='row'>
-                {row.name}
-              </TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </TableContainer>
-  );
-};
-
-export default ClassTable;
