+++ src/storybook/PaletteItem.js
@@ -1,9 +1,16 @@
-import Grid from '@mui/material/Grid';
 import React from 'react';
 
 const PaletteItem = ({ backgroundColor, classKey }) => {
   return (
-    <Grid item xs={12} md={4} display='flex' direction='row' alignItems='flex-start' justifyContent='flex-start'>
+    <div
+      style={{
+        display: 'flex',
+        alignItems: 'flex-start',
+        flexBasis: '33%',
+        flexDirection: 'row',
+        justifyContent: 'flex-start',
+      }}
+    >
       <div
         style={{
           color: 'transparent',
@@ -22,7 +29,7 @@ const PaletteItem = ({ backgroundColor, classKey }) => {
         <p style={{ margin: '0' }}>{classKey}</p>
         <b>{backgroundColor}</b>
       </div>
-    </Grid>
+    </div>
   );
 };
